requirements.txt
Django==1.9
django-extensions==1.7.2
django-filter==0.14.0
djangorestframework==3.4.4
requests==2.11.1
